Fixed gas limit multiplier method

+++ core/chains/evm/gas/arbitrum_estimator_test.go
@@ -109,7 +109,7 @@ func TestArbitrumEstimator(t *testing.T) {
 		require.Error(t, err)
 		assert.EqualError(t, err, "estimated gas price: 42 wei is greater than the maximum gas price configured: 40 wei")
 		assert.Nil(t, gasPrice)
-		assert.Equal(t, uint32(0), chainSpecificGasLimit)
+		assert.Equal(t, uint64(0), chainSpecificGasLimit)
 	})
 
 	t.Run("gas price is lower than global max gas price", func(t *testing.T) {
@@ -133,7 +133,7 @@ func TestArbitrumEstimator(t *testing.T) {
 		gasPrice, chainSpecificGasLimit, err := o.GetLegacyGas(testutils.Context(t), calldata, gasLimit, assets.NewWeiI(110))
 		assert.EqualError(t, err, "estimated gas price: 120 wei is greater than the maximum gas price configured: 110 wei")
 		assert.Nil(t, gasPrice)
-		assert.Equal(t, uint32(0), chainSpecificGasLimit)
+		assert.Equal(t, uint64(0), chainSpecificGasLimit)
 	})
 
 	t.Run("calling BumpLegacyGas on unstarted arbitrum estimator returns error", func(t *testing.T) {
